backend/requirements.txt
# Mini-OpenClaw Backend Dependencies
# Python 3.12 recommended

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Agent Orchestration
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# LangChain Community & Experimental
langchain-community>=0.3.0
langchain-experimental>=0.3.0

# RAG Engine
llama-index>=0.12.0
llama-index-core>=0.12.0
llama-index-retrievers-bm25>=0.4.0
llama-index-embeddings-openai>=0.3.0

# Neo4j Graph Database
neo4j>=5.0.0

# Model Interface
openai>=1.54.0
httpx>=0.27.0
langchain-anthropic>=0.3.0

# Utilities
python-multipart>=0.0.12
pydantic>=2.9.0
pydantic-settings>=2.5.0
beautifulsoup4>=4.12.3
html2text>=2024.2.25
python-dotenv>=1.0.1
sse-starlette>=2.1.0
python-frontmatter>=1.1.0
tiktoken>=0.7.0
